Add useGoBack hook and refactor blog page
